Changed amount data types to decimal, refactored AddTransaction

+++ RtgsGlobal.TechTest.Api/Controllers/AccountController.cs
@@ -6,48 +6,48 @@ namespace RtgsGlobal.TechTest.Api.Controllers;
 [Route("[controller]")]
 public class AccountController : ControllerBase
 {
-	private readonly IAccountProvider _accountProvider;
+    private readonly IAccountProvider _accountProvider;
 
-	public AccountController(IAccountProvider accountProvider)
-	{
-		_accountProvider = accountProvider;
-	}
+    public AccountController(IAccountProvider accountProvider)
+    {
+        _accountProvider = accountProvider;
+    }
 
-	[HttpPost("{accountIdentifier}", Name = "Deposit")]
-	public IActionResult Deposit(string accountIdentifier, [FromBody] float amount)
-	{
-		_accountProvider.Deposit(accountIdentifier, amount);
-		return Ok();
-	}
+    [HttpPost("{accountIdentifier}", Name = "Deposit")]
+    public IActionResult Deposit(string accountIdentifier, [FromBody] decimal amount)
+    {
+        _accountProvider.Deposit(accountIdentifier, amount);
+        return Ok();
+    }
 
-	[HttpPost("{accountIdentifier}/withdraw", Name = "Withdrawal")]
-	public IActionResult Withdraw(string accountIdentifier, [FromBody] float amount)
-	{
-		_accountProvider.Withdraw(accountIdentifier, amount);
-		return Ok();
-	}
+    [HttpPost("{accountIdentifier}/withdraw", Name = "Withdrawal")]
+    public IActionResult Withdraw(string accountIdentifier, [FromBody] decimal amount)
+    {
+        _accountProvider.Withdraw(accountIdentifier, amount);
+        return Ok();
+    }
 
-	[HttpPost("transfer", Name = "Transfer")]
-	public IActionResult Transfer(MyTransferDto transfer)
-	{
-		_accountProvider.Transfer(transfer);
-		return Accepted();
-	}
+    [HttpPost("transfer", Name = "Transfer")]
+    public IActionResult Transfer(MyTransferDto transfer)
+    {
+        _accountProvider.Transfer(transfer);
+        return Accepted();
+    }
 
-	[HttpGet("{accountIdentifier}", Name = "GetBalance")]
-	public MyBalance Get(string accountIdentifier) => _accountProvider.GetBalance(accountIdentifier);
+    [HttpGet("{accountIdentifier}", Name = "GetBalance")]
+    public MyBalance Get(string accountIdentifier) => _accountProvider.GetBalance(accountIdentifier);
 }
 
 public class MyTransferDto
 {
-	public MyTransferDto(string debtorAccountIdentifier, string creditorAccountIdentifier, float amount)
-	{
-		DebtorAccountIdentifier = debtorAccountIdentifier;
-		CreditorAccountIdentifier = creditorAccountIdentifier;
-		Amount = amount;
-	}
+    public MyTransferDto(string debtorAccountIdentifier, string creditorAccountIdentifier, decimal amount)
+    {
+        DebtorAccountIdentifier = debtorAccountIdentifier;
+        CreditorAccountIdentifier = creditorAccountIdentifier;
+        Amount = amount;
+    }
 
-	public string DebtorAccountIdentifier { get; set; }
-	public string CreditorAccountIdentifier { get; set; }
-	public float Amount { get; set; }
+    public string DebtorAccountIdentifier { get; set; }
+    public string CreditorAccountIdentifier { get; set; }
+    public decimal Amount { get; set; }
 }
